multi-select web component for adding/removing tags
There are many ways of solving this and I went back and forth quite a
bit trying to find a simple solution that feels right.

For now I'm settling with _not using a shadowroot_, which simplifies a
lot of things quite a bit (e.g. form submission, style inheritance),
while having few drawbacks, like having to use classes instead of ids
for query selecting, because of possible id clashes.

The solution is not yet completely finished as the component requires a
lot of html. So ideally we would want to turn this into a server
generated component in the next step.